Refactor out quickpick to util
